limit the public API
